+++ src/helpers/ApiHelper.php
@@ -1,0 +1,20 @@
+<?php
+
+
+namespace escape\escapedam\helpers;
+
+use Craft;
+
+/**
+ * Class ApiHelper
+ * @package escape\escapedam\helpers
+ */
+class ApiHelper
+{
+
+    /** @var \string[][] */
+    const LANGUAGE_CODE_MAP = [
+        'nb' => ['nb-NO', 'nn', 'nn-NO', 'no'],
+    ];
+
+}